+++ platform/flowglad-next/src/server/index.ts
@@ -37,6 +37,8 @@ import { catalogsRouter } from './routers/catalogsRouter'
 import { usageMetersRouter } from './routers/usageMetersRouter'
 import { usageEventsRouter } from './routers/usageEventsRouter'
 import { inviteUserToOrganization } from './mutations/inviteUserToOrganization'
+import { t } from './coreTrpcObject'
+
 const purchasesRouter = router({
   create: createPurchase,
   update: editPurchase,