+++ platform/flowglad-next/src/utils/stripe.ts
@@ -1171,8 +1171,6 @@ export const getStripePrice = async (
  * meaning to create an intent for an anonymized customer to create a subscription.
  */
 export const createSetupIntentForCheckoutSession = async (params: {
-  price: Price.SubscriptionRecord | Price.UsageRecord
-  product: Product.Record
   organization: Organization.Record
   checkoutSession: CheckoutSession.Record
   purchase?: Purchase.Record