+++ packages/express/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @flowglad/express
 
+## 0.10.0
+
+### Minor Changes
+
+- Add subscription method, many other improvements
+
+### Patch Changes
+
+- Updated dependencies
+  - @flowglad/server@0.10.0
+  - @flowglad/shared@0.10.0
+  - @flowglad/react@0.10.0
+
 ## 0.9.1
 
 ### Patch Changes